+++ src/Command/CreatePluginsLinksCommand.php
@@ -1,0 +1,51 @@
+<?php
+/**
+ * This file is part of me-tools.
+ *
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/me-tools
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
+ * @since       2.18.0
+ */
+namespace MeTools\Command;
+
+use Cake\Console\Arguments;
+use Cake\Console\ConsoleIo;
+use Cake\Console\ConsoleOptionParser;
+use Cake\Shell\Task\AssetsTask;
+use MeTools\Console\Command;
+
+/**
+ * 'Creates symbolic links for plugins assets'
+ */
+class CreatePluginsLinksCommand extends Command
+{
+    /**
+     * Hook method for defining this command's option parser
+     * @param ConsoleOptionParser $parser The parser to be defined
+     * @return ConsoleOptionParser
+     */
+    protected function buildOptionParser(ConsoleOptionParser $parser)
+    {
+        $parser->setDescription(__d('me_tools', 'Creates symbolic links for plugins assets'));
+
+        return $parser;
+    }
+
+    /**
+     * 'Creates symbolic links for plugins assets'
+     * @param Arguments $args The command arguments
+     * @param ConsoleIo $io The console io
+     * @return null|int The exit code or null for success
+     */
+    public function execute(Arguments $args, ConsoleIo $io)
+    {
+        (new AssetsTask($io))->symlink();
+
+        return null;
+    }
+}